+++ sdk/core/azure-core/src/uuid.cpp
@@ -3,8 +3,15 @@
 
 #include "azure/core/uuid.hpp"
 
+#include "azure/core/internal/strings.hpp"
+#include "azure/core/platform.hpp"
+
+#include <cassert>
 #include <cstdio>
+#include <cstring>
 #include <random>
+#include <stdexcept>
+#include <type_traits>
 
 #if defined(AZ_PLATFORM_POSIX)
 #include <thread>
@@ -16,78 +23,159 @@ static thread_local std::mt19937_64 randomGenerator(std::random_device{}());
 } // namespace
 #endif
 
+using Azure::Core::_internal::StringExtensions;
+
+namespace {
+/*
+"00000000-0000-0000-0000-000000000000"
+         ^    ^    ^    ^
+ 000000000011111111112222222222333333
+ 012345678901234567890123456789012345
+ \______________ = 36 ______________/
+*/
+constexpr size_t UuidStringLength = 36;
+constexpr bool IsDashIndex(size_t i) { return i == 8 || i == 13 || i == 18 || i == 23; }
+
+constexpr std::uint8_t HexToNibble(char c) // does not check for errors
+{
+  if (c >= 'a')
+  {
+    return 10 + (c - 'a');
+  }
+
+  if (c >= 'A')
+  {
+    return 10 + (c - 'A');
+  }
+
+  return c - '0';
+}
+
+constexpr char NibbleToHex(std::uint8_t nibble) // does not check for errors
+{
+  if (nibble <= 9)
+  {
+    return '0' + nibble;
+  }
+
+  return 'a' + (nibble - 10);
+}
+} // namespace
+
 namespace Azure { namespace Core {
   std::string Uuid::ToString() const
   {
-    // Guid is 36 characters
-    //  Add one byte for the \0
-    char s[37];
+    std::string s(UuidStringLength, '-');
 
-    std::snprintf(
-        s,
-        sizeof(s),
-        "%2.2x%2.2x%2.2x%2.2x-%2.2x%2.2x-%2.2x%2.2x-%2.2x%2.2x-%2.2x%2.2x%2.2x%2.2x%2.2x%2.2x",
-        m_uuid[0],
-        m_uuid[1],
-        m_uuid[2],
-        m_uuid[3],
-        m_uuid[4],
-        m_uuid[5],
-        m_uuid[6],
-        m_uuid[7],
-        m_uuid[8],
-        m_uuid[9],
-        m_uuid[10],
-        m_uuid[11],
-        m_uuid[12],
-        m_uuid[13],
-        m_uuid[14],
-        m_uuid[15]);
+    for (size_t bi = 0, si = 0; bi < m_uuid.size(); ++bi)
+    {
+      if (IsDashIndex(si))
+      {
+        ++si;
+      }
 
-    return std::string(s);
+      assert((si < UuidStringLength) && (si + 1 < UuidStringLength));
+
+      const std::uint8_t b = m_uuid[bi];
+      s[si] = NibbleToHex((b >> 4) & 0x0F);
+      s[si + 1] = NibbleToHex(b & 0x0F);
+      si += 2;
+    }
+
+    return s;
+  }
+
+  Uuid Uuid::Parse(std::string const& uuidString)
+  {
+    bool parseError = false;
+    Uuid result;
+    if (uuidString.size() != UuidStringLength)
+    {
+      parseError = true;
+    }
+    else
+    {
+      // si = string index, bi = byte index
+      for (size_t si = 0, bi = 0; si < UuidStringLength; ++si)
+      {
+        const auto c = uuidString[si];
+        if (IsDashIndex(si))
+        {
+          if (c != '-')
+          {
+            parseError = true;
+            break;
+          }
+        }
+        else
+        {
+          assert(si + 1 < UuidStringLength && bi < result.m_uuid.size());
+
+          const auto c2 = uuidString[si + 1];
+          if (!StringExtensions::IsHexDigit(c) || !StringExtensions::IsHexDigit(c2))
+          {
+            parseError = true;
+            break;
+          }
+
+          result.m_uuid[bi] = (HexToNibble(c) << 4) | HexToNibble(c2);
+          ++si;
+          ++bi;
+        }
+      }
+    }
+
+    return parseError ? throw std::invalid_argument(
+               "Error parsing Uuid: '" + uuidString
+               + "' is not in the '00112233-4455-6677-8899-aAbBcCdDeEfF' format.")
+                      : result;
   }
 
   Uuid Uuid::CreateUuid()
   {
-    uint8_t uuid[UuidSize] = {};
+    Uuid result{};
+
+    // Using RngResultType and RngResultSize to highlight the places where the same type is used.
+    using RngResultType = std::uint32_t;
+    static_assert(sizeof(RngResultType) == 4, "sizeof(RngResultType) must be 4.");
+    constexpr size_t RngResultSize = 4;
 
 #if defined(AZ_PLATFORM_WINDOWS)
     std::random_device rd;
+
+    static_assert(
+        std::is_same<RngResultType, decltype(rd())>::value,
+        "random_device::result_type must be of RngResultType.");
 #else
-    std::uniform_int_distribution<uint32_t> distribution;
+    std::uniform_int_distribution<RngResultType> distribution;
 #endif
 
-    for (size_t i = 0; i < UuidSize; i += 4)
+    for (size_t i = 0; i < result.m_uuid.size(); i += RngResultSize)
     {
 #if defined(AZ_PLATFORM_WINDOWS)
-      const uint32_t x = rd();
+      const RngResultType x = rd();
 #else
-      const uint32_t x = distribution(randomGenerator);
+      const RngResultType x = distribution(randomGenerator);
 #endif
-      std::memcpy(uuid + i, &x, 4);
+      std::memcpy(result.m_uuid.data() + i, &x, RngResultSize);
     }
 
     // The variant field consists of a variable number of the most significant bits of octet 8 of
     // the UUID.
-    // https://www.rfc-editor.org/rfc/rfc4122.html#section-4.1.1
-    // For setting the variant to conform to RFC4122, the high bits need to be of the form 10xx,
+    // https://www.rfc-editor.org/rfc/rfc9562.html#name-variant-field
+    // For setting the variant to conform to RFC9562, the high bits need to be of the form 10xx,
     // which means the hex value of the first 4 bits can only be either 8, 9, A|a, B|b. The 0-7
     // values are reserved for backward compatibility. The C|c, D|d values are reserved for
     // Microsoft, and the E|e, F|f values are reserved for future use.
     // Therefore, we have to zero out the two high bits, and then set the highest bit to 1.
-    uuid[8] = (uuid[8] & 0x3F) | 0x80;
+    result.m_uuid.data()[8] = (result.m_uuid.data()[8] & 0x3F) | 0x80;
 
-    constexpr uint8_t version = 4; // Version 4: Pseudo-random number
+    {
+      // https://www.rfc-editor.org/rfc/rfc9562.html#name-version-field
+      constexpr std::uint8_t Version = 4; // Version 4: Pseudo-random number
+      result.m_uuid.data()[6] = (result.m_uuid.data()[6] & 0xF) | (Version << 4);
+    }
 
-    uuid[6] = (uuid[6] & 0xF) | (version << 4);
-
-    return Uuid(uuid);
+    return result;
   }
-
-  Uuid Uuid::CreateFromArray(std::array<uint8_t, UuidSize> const& uuid)
-  {
-    Uuid rv{uuid.data()};
-    return rv;
-  }
-
 }} // namespace Azure::Core